refactor(cli): extract scaffold package spec resolver
